// src/domain/systems/supply/constants/supplySystem.enums.ts
export const enum ESupplySystemType {
  AIR_HANDLING_UNIT = "air_handling_unit", // 공조기
  ELECTRIC_RADIATOR = "electric_radiator", // 전기 라디에이터
  FAN_COIL_UNIT = "fan_coil_unit", // 팬코일유닛
  PACKAGED_AIR_CONDITIONER = "packaged_air_conditioner", // 패키지 에어컨
  RADIANT_FLOOR = "radiant_floor", // 온돌/복사 바닥
  RADIATOR = "radiator", // 라디에이터
}